requirements.txt
streamlit==1.39.0
google-generativeai==0.8.3
pandas==2.2.3
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.2
openpyxl==3.1.5
plotly==5.24.1
altair==5.4.1
